+++ packages/golden_toolkit/lib/src/golden_builder/golden_builder_table.dart
@@ -3,49 +3,57 @@ part of 'golden_builder.dart';
 typedef TableColumnWidthMap = Map<int, TableColumnWidth>;
 
 class GoldenBuilderTable extends GoldenBuilder {
+  /// Helper to create flexible columns with equal width distribution instead of the default [IntrinsicColumnWidth].
   static TableColumnWidthMap flexibleColumns(int count) {
     return {for (var index in List.generate(count, (index) => index)) index: const FlexColumnWidth(1)};
   }
 
-  static TableColumnWidthMap fixedColumns({required int count, required double columnWidth}) {
-    return {for (var index in List.generate(count, (index) => index)) index: FixedColumnWidth(columnWidth)};
+  /// Helper to create fixed width columns instead of the default [IntrinsicColumnWidth].
+  static TableColumnWidthMap fixedColumns({required int count, required double width}) {
+    return {for (var index in List.generate(count, (index) => index)) index: FixedColumnWidth(width)};
   }
 
   GoldenBuilderTable({
     required this.columns,
-    this.rowSpacing = 16.0,
+    this.columnSpacing = 16.0,
+    this.columnWidths = const {},
     super.wrap,
     super.bgColor,
-    this.columnWidths = const {},
   });
 
   /// number of columns to distribute the scenarios in
   final int columns;
 
   /// horizontal spacing between columns
-  final double rowSpacing;
+  /// defaults to [16.0]
+  final double columnSpacing;
 
   /// column widths for the actual content columns
-  /// if a column width is not provided, it will default to IntrinsicColumnWidth
-  /// GoldenBuilderTable.flexibleColumns() and GoldenBuilderTable.fixedColumns() are available, as helpers.
+  ///
+  /// if a column width is not provided, it will default to [IntrinsicColumnWidth]
+  ///
+  /// [GoldenBuilderTable.flexibleColumns()] and [GoldenBuilderTable.fixedColumns()] are available, as helpers.
   final TableColumnWidthMap columnWidths;
 
   int get _effectiveRows => (scenarios.length / columns).ceil();
 
   /// for example:
   /// columns = 2
+  ///
   /// columnWidths expects the user to provide like {0: FlexColumnWidth(1), 1: FlexColumnWidth(1)}
+  ///
   /// but in reality, it has 3 columns, because of the spacing column, so the effective
-  /// columnWidths will be {0: FlexColumnWidth(1), 1: FixedColumnWidth(rowSpacing), 2: FlexColumnWidth(1)}
+  /// columnWidths will be {0: FlexColumnWidth(1), 1: FixedColumnWidth(columnSpacing), 2: FlexColumnWidth(1)}
+  ///
   /// if a column width is not provided, it will default to IntrinsicColumnWidth
-  /// spacing columns are always FixedColumnWidth(rowSpacing)
+  /// spacing columns are always FixedColumnWidth(columnSpacing)
   TableColumnWidthMap _effectiveColumnWidths() {
     final widths = <int, TableColumnWidth>{};
     for (int i = 0; i < columns; i++) {
       widths[i * 2] = columnWidths[i] ?? const IntrinsicColumnWidth();
     }
     for (int i = 0; i < columns - 1; i++) {
-      widths[i * 2 + 1] = FixedColumnWidth(rowSpacing);
+      widths[i * 2 + 1] = FixedColumnWidth(columnSpacing);
     }
     return widths;
   }