+++ packages/golden_toolkit/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.16.5
 
-- Deprecate GoldenBuilder.grid() in favor of GoldenBuilder.table()
+- Deprecate GoldenBuilder.column() in favor of GoldenBuilderColumn()
+- Deprecate GoldenBuilder.grid() in favor of GoldenBuilderTable()
+- Add flexibleColumns() and fixedColumns() as static helper methods on GoldenBuilderTable for easier configuration of column widths
+- Rename rowSpacing to columnSpacing
 
 ## 0.16.4